feat: adding to save promocodes to localstorage
