+++ src/components/cart/index.ts
@@ -1,36 +1,18 @@
 import Component from '../../common/component';
 import Basket from '../../common/basket';
-import { IPromo, IProduct, ICart } from '../../common/interface';
+import { IProduct, ICart, IPromo } from '../../common/interface';
+import { validPromo } from '../../common/basket';
 import './style-cart.scss';
 
 
-const validPromo: IPromo[] = [
-  {
-    promoname: 'NEWYEAR',
-    description: 'Happy New Year',
-    discount: 23
-  },
-  {
-    promoname: 'RS',
-    description: 'Rolling Scopes School',
-    discount: 10
-  },
-  {
-    promoname: 'MC',
-    description: 'Merry Christmas',
-    discount: 7
-  },
-]
-
 export default class Cart extends Component {
   private basket: Basket;
-  private promocodes: IPromo[];
+  
   private cartParams: ICart;
 
   constructor(name: string) {
     super(name);
     this.basket = new Basket();
-    this.promocodes = [];
     const cartSave:string | null = localStorage.getItem('cartParams');
     if (cartSave) {
       this.cartParams = JSON.parse(cartSave);
@@ -51,6 +33,7 @@ export default class Cart extends Component {
     blockProducts.style.height = `${Math.min(this.cartParams.maxItems, this.basket.getTotalProducts()) * 100}px`;
 
     this.draw();
+    this.drawPromoCodes();
     this.drawSummary();
     this.basket.drawHeader();
     this.initEvents();
@@ -386,11 +369,6 @@ export default class Cart extends Component {
     return `?page=${page}`;
   }
 
-  private getPromoCode(name: string): IPromo {
-    const index: number = validPromo.findIndex(el => el.promoname === name);
-    return validPromo[index];
-  }
-
   private findPromoCode(e: Event): void {
     const target = e.target as HTMLInputElement;
     const index: number = validPromo.findIndex(el => el.promoname === target.value);
@@ -406,7 +384,7 @@ export default class Cart extends Component {
 
   private addPromoCode(): void {
     const promoName = document.querySelector('.cart-summary__promo') as HTMLInputElement;
-    this.promocodes.push(this.getPromoCode(promoName.value));
+    this.basket.addPromoCode(promoName.value);
     this.drawPromoCodes();
     promoName.value = '';
     promoName.dispatchEvent(new Event('input', { bubbles: true }));
@@ -418,8 +396,7 @@ export default class Cart extends Component {
     const codeName = liItem.dataset.promoname;
     
     if (codeName) {
-      const index = this.getPromoCodesIndex(codeName);
-      this.promocodes.splice(index, 1);
+      this.basket.deletePromoCode(codeName);
       liItem?.parentNode?.removeChild(liItem);
       this.drawPromoCodes();
     }
@@ -430,8 +407,8 @@ export default class Cart extends Component {
     const promoItems = document.querySelector('.cart-summary__promo-items') as HTMLElement;
     const promoTotal = document.querySelector('.cart-summary__total-promo') as HTMLElement;
     promoItems.innerHTML = '';
-    if (this.promocodes.length > 0) {
-      this.promocodes.forEach((element: IPromo) => {
+    if (this.basket.promocodes.length > 0) {
+      this.basket.promocodes.forEach((element: IPromo) => {
         const liItem = document.createElement('li');
         const liDrop = document.createElement('button');
         liItem.className = 'cart-summary__promo-item';
@@ -459,8 +436,8 @@ export default class Cart extends Component {
     let promoSumm = 0;
     const totalSumm = this.basket.getTotalSumm();
     let prevSumm = totalSumm;
-    if (this.promocodes.length > 0) {
-      this.promocodes.forEach((element: IPromo) => {
+    if (this.basket.promocodes.length > 0) {
+      this.basket.promocodes.forEach((element: IPromo) => {
         promoSumm = Math.round((prevSumm * (100 - element.discount) / 100) * 100) / 100;
         prevSumm = promoSumm;
       });
@@ -469,16 +446,6 @@ export default class Cart extends Component {
     } else {
       summaryTotalSumm.textContent = `${totalSumm.toFixed(2)}$`;
     }
-  }
-
-  private getPromoCodesIndex(name: string): number {
-    let result = -1;
-    for (let i = 0; i < this.promocodes.length; i++) {
-      if (this.promocodes[i].promoname === name) {
-        result = i;
-      }
-    }
-    return result;
   }
 
   public setToLocalStorage(): void {
